flashcards.java
Flashcard application. Users can add/remove flashcards using this. If the users wish to be tested based on the flashcard that is also possible with the program. The application allows simple file i/o by allowing the user to import or export the flashcard to and from a text file that is saved within the same directory.